Add physics
* XML: Support self-closing tags

* add Physics and Bugs

* doesn't work but different 

but works in debugger => timing?

* fix coroutine lambda object lifetime bug

coroutine lifetime > lambda object (captures) lifetime => not good

solution: sucky shared_ptr<void> and static stuff :c

* change velocity on collision

* add Documentation

+++ zoe/src/zoe/display/Camera.h
@@ -133,7 +133,7 @@ public:
      * @param width the specified camera width
      * @param height the specified camera height
      */
-    Camera2D(const vec2& position, const float& width, const float& height);
+    Camera2D(const vec2& position, const double& width, const double& height);
 
     /**
      * Create an empty Camera2D.
@@ -153,7 +153,7 @@ protected:
     void fill(const XMLNode &node) override;
 
 private:
-    float width, height;
+    double width, height;
 };
 
 /**
@@ -168,7 +168,7 @@ public:
      * @param fov the specified field of view. The Fov is given in degrees
      * @param ratio the specified ration. It can be calculated by width/height
      */
-    Camera3D(const vec3& position, const vec3& rotation, const float& fov, const float& ratio);
+    Camera3D(const vec3& position, const vec3& rotation, const double& fov, const double& ratio);
 
     /**
      * Create an empty Camera3D.
@@ -185,19 +185,19 @@ public:
      * Set the pitch of the Camera. The view matrix and inverse view matrix are updated.
      * @param pitch the new camera pitch in radians
      */
-    void setPitch(const float& pitch);
+    void setPitch(const double& pitch);
 
     /**
      * Set the yaw of the Camera. The view matrix and inverse view matrix are updated.
      * @param yaw the new camera yaw in radians
      */
-    void setYaw(const float& yaw);
+    void setYaw(const double& yaw);
 
     /**
      * Set the roll of the Camera. The view matrix and inverse view matrix are updated.
      * @param roll the new camera roll in radians
      */
-    void setRoll(const float& roll);
+    void setRoll(const double& roll);
 
     /**
      * Set the rotation of the Camera to the specified pitch, yaw and roll. This function equals `Camera::setRotation(vec3({pitch, yaw, roll}))`.
@@ -206,7 +206,7 @@ public:
      * @param yaw the specified yaw in radians
      * @param roll the specified roll in radians
      */
-    void setRotation(const float& pitch, const float& yaw, const float& roll);
+    void setRotation(const double& pitch, const double& yaw, const double& roll);
 
     /**
      * Set the direction that the camera is pointing.
@@ -221,21 +221,21 @@ public:
      * @param direction the new direction
      * @param roll the new roll
      */
-    void setDirection(const vec3& direction, const float& roll);
+    void setDirection(const vec3& direction, const double& roll);
 
     /**
      * Set the field of view of the camera.
      * Thew view matrix and inverse view matrix are updated.
      * @param fov the fov in degrees.
      */
-    void setFoV(const float& fov);
+    void setFoV(const double& fov);
 
     /**
      * Set the ratio of width and height. Can be calculated by width/height.
      * Thew view matrix and inverse view matrix are updated.
      * @param ratio the width and height
      */
-    void setRatio(const float& ratio);
+    void setRatio(const double& ratio);
 
     /**
      * Rotates this camera by the given value in radians.
@@ -260,12 +260,12 @@ public:
      * Returns the field of view of the Camera.
      * @return the FoV in degrees
      */
-    [[nodiscard]] inline const float& getFoV() const { return fov; }
+    [[nodiscard]] inline const double& getFoV() const { return fov; }
     /**
      * Returns the ratio of width and height. Can be calculated by `width/height`
      * @return the ratio
      */
-    [[nodiscard]] inline const float& getRatio() const { return ratio; }
+    [[nodiscard]] inline const double& getRatio() const { return ratio; }
 protected:
 
     /**
@@ -281,7 +281,7 @@ protected:
 
 private:
     vec3 rotation;
-    float fov, ratio;
+    double fov, ratio;
 };
 
 }
